+++ cmd/timetracker/cmd/tray.go
@@ -5,6 +5,7 @@ import (
 	"github.com/neflyte/timetracker/internal/appstate"
 	"github.com/neflyte/timetracker/internal/logger"
 	"github.com/neflyte/timetracker/internal/ui/tray"
+	"github.com/neflyte/timetracker/internal/utils"
 	"github.com/nightlyone/lockfile"
 	"github.com/spf13/cobra"
 	"os"
@@ -44,12 +45,22 @@ func preDoTray(_ *cobra.Command, _ []string) error {
 	}
 	trayCmdLockfilePath = path.Join(userConfigDir, trayPidfile)
 	log.Trace().Msgf("trayCmdLockfilePath=%s", trayCmdLockfilePath)
-	lockfileInfo, err := os.Stat(trayCmdLockfilePath)
-	if err != nil && !errors.Is(err, os.ErrNotExist) {
-		log.Err(err).Msgf("error stating pidfile %s", trayCmdLockfilePath)
+	pidExists, err := utils.CheckPidfile(trayCmdLockfilePath)
+	if err != nil && !errors.Is(err, os.ErrNotExist) && !errors.Is(err, utils.ErrStalePidfile) {
+		log.Err(err).Msgf("error checking pidfile %s", trayCmdLockfilePath)
 	}
-	if lockfileInfo != nil {
-		log.Warn().Msgf("pidfile %s exists; modtime=%s", trayCmdLockfilePath, lockfileInfo.ModTime().String())
+	log.Trace().Msgf("pidExists=%t", pidExists)
+	if pidExists {
+		log.Error().Msgf("pidfile %s exists and its process is running; exiting", trayCmdLockfilePath)
+		return errors.New("another process is already running")
+	}
+	if errors.Is(err, utils.ErrStalePidfile) {
+		log.Debug().Msgf("attempting to remove stale pidfile %s", trayCmdLockfilePath)
+		err = os.Remove(trayCmdLockfilePath)
+		if err != nil && !errors.Is(err, os.ErrNotExist) {
+			log.Err(err).Msgf("error removing existing pidfile %s", trayCmdLockfilePath)
+			return errors.New("unable to remove stale pidfile")
+		}
 	}
 	trayCmdLockFile, err = lockfile.New(trayCmdLockfilePath)
 	if err != nil {
@@ -78,8 +89,6 @@ func postDoTray(_ *cobra.Command, _ []string) error {
 		return err
 	}
 	log.Debug().Msgf("unlocked pidfile %s", trayCmdLockfilePath)
-	// Force-remove the pid file in case it didn't delete for some reason; we don't care if it fails
-	_ = os.Remove(trayCmdLockfilePath)
 	return nil
 }